Remove legacy E2eMessageSharedIcon component
